Modfication to the parallel framework in an attempt to track down the problem with multiple RHSs. It appears to be working again in serial mode, but parallel mode runs, but the answer is garbage

+++ src/grid.F90
@@ -468,7 +468,7 @@ contains
         implicit none
 
         type(gridBlock), intent(inout) :: gAll(:)
-        integer(kind=long), intent(in) :: nPts_tot
+        integer, intent(in) :: nPts_tot
 
         integer :: wMe, iMe, jMe, me, nbr, offSet, label
